Move Value, push aggregation into it
Okay, as a part of #290 we're going to have to have a concept of
binning. Right now, all Value are wrapper either around an f64 or a
CKMS. We need to break that up. This is the start of that. I _think_
I can keep the individual changes small based on some ineffective
giant changes I've been making all day.

Telemetry is the hardest part of cernan. I can't tell if it's overly
clever or just as clever as its demands make it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>