Push Aggregation down into Value  (#291)
* Move Value, push aggregation into it

Okay, as a part of #290 we're going to have to have a concept of
binning. Right now, all Value are wrapper either around an f64 or a
CKMS. We need to break that up. This is the start of that. I _think_
I can keep the individual changes small based on some ineffective
giant changes I've been making all day.

Telemetry is the hardest part of cernan. I can't tell if it's overly
clever or just as clever as its demands make it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Now passing tests

This commit fiddles with tests so that they now pass. There's some
commented out code but we'll fix that up, especially in buckets.rs.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* All existing tests pass

This commit fiddles with Value and Telemetry such that the tests
which are not commented out pass. The main action here is in Value.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Get test_all_insertions back online

This commit re-enables the last major test that had been commented
out. I _believe_ this means the change I've been making is ready
for review and comment.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>