Integrate quantiles/Histogram
This commit is the start to the integration of quantiles' Histogram.
It's big and painful and does not work yet. You'll notice that we're
pegged on an in-progress version of quantiles and that's needed while
we figure out what cernan needs exactly.

Heh, this one is painful to write. Telemetry is probably the worst
part of cernan to change, second to Bucket.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>