Another day, another block of changes to Telemetry
This commit introduces fixes for many of the operations on Telemetry
and resolves some of the test failures. We wrap Value in an Option
not because there might be a None in a valid state but so that we
can avoid allocing another large structure when we need to do
AddAssign. Annoying.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>