More fiddling with zero padding
There are still low-probability conditions that we fail on.
This commit addresses the ones I've found so far by making
sure we feed last seen through, allowing us, as the comments
note, to pad into the past. This commit suffers from out-of-order
delivery which should not be possible. I think this may have
existed prior to this commit, but I'm unsure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>