Finish off padding tests
There's still a known issue with this implementation but it's a
deeper problem. Namely this: if you have two points with the same
timestamp arrive into a sink across a flush you'll get two distinct
points, rather than one unified, aggregated point. This is not
addressable in the context of this work and will be a follow-on
project.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>