Introduce 'delay_filter' (#296)
This is the first hard-coded filter introduced into cernan. The
intention is that the filter will drop any metric::Event which
comes through and is too old, as configured by the filter's
tolerance.

I've changed the version in Cargo.toml to 0.7.0-pre as the config
file changes are not backward compatible with 0.6.x. We'll have to
update the wiki separately.

Related to #295

Signed-off-by: Brian L. Troutwine <blt@postmates.com>