Add Documentation, Moves to Quiet Clippy  (#298)
* Add documentation, clippy changes

This commit touches a good chunk of the codebase. We now have docs
required enabled and are quiet against clippy. Well, quiet except
in that rustc warns about the clippy directives it can't understand.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Fix TagMap PartialEq implementation

This commit corrects the explicit PartialEq implementation. There
are still a scant few failing tests.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Correct tests, clippy runs

Just fiddling at this point. Doc tests are the hardest to get right
as their feedback cycle is so long.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>