Move tagmap to use string caching
The TagMaps inside LogLine and Telemetry now use string caching for
their owned strings. This will reduce memory requirements for maps
which use more than eight characters in a key or value. The downside
is increased read / write time to maps owing to the need to lock on
the cache but, you know, that's the usual tradeoff I guess.

I would like to cache the whole of the TagMap but we've stuffed
it into an Arc. That'll take a while to untangle.

I'm not super happy with the API here, especially iteration. It
might make more sent to provide an iterator over &str, rather than
hash id. Current approach forces continual lookups in client-code.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>