Move string cache to use SeaHasher
The default HashMap in rust is pretty pokey. This commit moves
the string cache -- which needs to be not pokey -- away from
the default hasher. This cuts down on the CPU requirement
somewhat.

I'm not totally satisfied with the performance characteristics
of cernan these days. We've added quite a bit of capability but,
at the same time, we've traded off in terms of CPU.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>