Simplify Buckets iteration, correct missing metrics
This commit corrects the issue with cernan that started off the whole
chain of investigation. I'm not 100% sure what the issue was, other
than that the indexing present in the old implementation of iteration
for buckets::Iter was too complicated and gave wrong results. While
simplifying the implementation to understand it better I also fixed
what ailed cernan.

Welp.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>