Bump to 0.7.1
This is a very minor change. The flush_boundary filter no longer
has a valve in place, of which was not feasible to have on a
higher-load system. There's some thinking that needs to be done
to manage overload conditions via the valve systems for end
users. This commit is not that thinking.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>