Padding control for wavefront sink
This commit adds configuration knobs in the wavefront sink to
turn zero-padding on and off by aggregation type. Padding was
introduced in 0.7.0 but there's been concern with the results
now that 0.7.1 is out and about. We now allow users to decide to
pad or not.

By default this feature is OFF. Research by @pulltab suggests
that neither etsy/statsd nor statsite have this feature.

This resolves #301

Signed-off-by: Brian L. Troutwine <blt@postmates.com>