Bump to 0.7.4
This release two primary changes to the project:

  * b2236f0 :: Introduce reading tags from environment variables
  * f7ced09 :: Drain unbounded buffer in Internal source

The first change is motivated by the needs of @blakebarnett for
use in k8s. It is hoped this change will ease the use of cernan
in that environment by allowing tag values to be read from, uh,
the environment. The second change removes a potential place of
unbounded growth in cernan.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>