Deallocate unused structures in cernan startup
This commit deallocates the unused structures that are created as
a part of the cernan startup routine. This is especially important
with regard to the hopper Senders present in 'senders' and 'flush_sends'
as these contain file descriptors which will NOT be closed until
such time as the Sender's original thread lets go of it.

This is related to the work being done as a part of #244.

This commit has the benefit is slightly reducing cernan's memory draw.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>