Drop deleted files from the file source
Prior to this commit the file source has assumed that descriptors
to deleted files can be cleared out when a new file is dropped in
the path. This is a problem for clients where log files are made
only at boot and then deleted. We now clear files in the same way
we used too but also when NotFound is returned from io operations.

This resolves #244

Signed-off-by: Brian L. Troutwine <blt@postmates.com>