Drop deleted files from the file source (#310)
* Drop deleted files from the file source

Prior to this commit the file source has assumed that descriptors
to deleted files can be cleared out when a new file is dropped in
the path. This is a problem for clients where log files are made
only at boot and then deleted. We now clear files in the same way
we used too but also when NotFound is returned from io operations.

This resolves #244

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Break up the file source

This commit breaks up the file source in a fashion which drops
FPs that are bound to paths which no longer exist, rotate faster
and with simpler code. Unfortunately we've changed how the file_watcher
read_line works and this has made the QC test for file_watcher sit
and spin. The next commit will have to correct that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Adapt file source qc test to new API

With the changes made to the file_watcher we broke the quickcheck
tests. Well! They're now fixed. Mostly we had to cope with the extra
O's that come out of the file_watcher.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>