Bump to cernan 0.7.5
This release contains fixes for the overlong holding of file
descriptors in a couple of spots inside of cernan:

  * 07d3fad0 :: drop unused structures after cernan boot
  * de747e12 :: drop deleted files from polling in file source

The first change allows hopper to drop file descriptors opened
as a part of cernan startup. The second change allows the file source
to drop file descriptors as soon as the files are marked deleted.
Previously we assumed that this could be avoided on the expectation
of applications creating a new logfile. This assumption was not
valid.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>