Avoid getting stuck on a closed sink valve (#313)
This commit corrects a condition where a sink can get stuck when
its valve state is closed. Consider the case where a flush is
required to clear a valve but no flush is possible because the valve
is closed. This is a deadlock.

This only affects bursty sinks with long flush times and low valve
thresholds, like the firehose sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>