Spruce up the cernan release build
We used to build releases inside Alpine Linux. This worked out okay for
a while but the rust version there has gotten very out of date as of
this writing. We're using ekidd/rust-musl-builder now, a Debian base
that'll crank out musl binaries. We're still targeting Alpine as the
deployment image.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>