Allow histograms to be flipped on for statsd metrics
This commit adds a 'mapping' capability for statsd source. It looks
like so:

    [sources]
      [sources.statsd.primary]
      enabled = true
      host = "localhost"
      port = 1024
      forwards = ["sinks.console", "sinks.null"]

      [sources.statsd.primary.mapping]
      [sources.statsd.primary.mapping.foo]
      mask = "foo.*"
      bounds = [0.0, 1.0, 10.0]

      [sources.statsd.primary.mapping.bar]
      mask = "bar.*"
      bounds = [0.0, 2.0, 20.0]

      [sources.statsd.primary.mapping.default]
      mask = "*"
      bounds = [0.0, 0.1, 10.0, 100.0]

In this example three mappings are created: 'foo', 'bar' and 'default'.
The 'foo' mapping will match those metrics whose names are 'foo.*' where
the glob is greedy, 'bar' matches 'bar.*' and default catches all that
remain. In this configuration no statsd timers or histograms will not
be an AggregationMethod::Histogram. If two mappings should overlap in
their masks the one which appears first in the configuration will be
the preferred mapping.

Resolves #307

Signed-off-by: Brian L. Troutwine <blt@postmates.com>