Correct prometheus sink QC test
The error here, though I suggested otherwise in the last commit
message, was legitimate. In particular, the find_match function
should not have searched both windowed and perpetual stores but
one or the other, depending on the AggregationMethod of the
input telemetry. That's because it's possible the model will have
made an aggregation with the same named metric of two distinct
aggregations.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>