Add HISTOGRAM support to the wavefront sink
This commit adds emission of the histogram summarization to the
wavefront sink. The name of the metric is smushed together with
the bin boundary and the count / timestamp happen where you'd
expect.

The bin is separated by '_' rather than a dot to avoid confusion
with bins that have a decimal value.

Resolves #308. Related to #307.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>