Remove flush loop, re-create client on each flush
If we fail to make one flush we assume that the client is ruined
and destroy it. In that case, there's no need to loop up to
attempt times in the ES flush. Also, because AWS ES will tend
to hang up if it's not contacted often enough we need to create
a new client each time anyhow.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>