Make Prometheus sink obey aggregations (#306)
* Make Prometheus sink obey aggregations

This commit makes the prometheus sink send a variety of types, rather
than the quantile, now that Telemetry is backed by different and
distinct kinds.

This is part of the work for #290.

This is related to the work of #285.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Allow prometheus values to fall off the end of the world

This commit makes it so that we only keep the last retain_limit
reported values. We never clear values unless more values come in
and push the older ones out.

Resolves #285

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update to master

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct plaintext, address comment issue

This commit addresses problems with the plaintext implementation
and removes a faulty comment in the source code. Nothing has changed
on the binary reporting side and I've confirmed that prometheus is
happy with both.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add commentary

This commit adds commentary to address Charles' questions about
the way in which aggregated points are retained and what mapping
exists between cernan's internal model and prometheus.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Absorb clippy suggestions

This commit incorporates clippy suggestions, excepting the
complexity on write_text which we simply allow.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Adjust documentation in the sink

We still referred to ten values per timestamp in the documentation
even when it should be retain_limit values per time series. This
is now corrected.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Split prometheus aggregation into perpetual / windowed

This commit does an initial pass on an implementation of windowed
and perpetual aggregations according to the conversation here:

    https://github.com/postmates/cernan/pull/306#discussion_r139770087

The QC tests are failing now because the model is quite different,
but that's okay and I'll chase this commit with a different
model.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct prometheus sink QC test

The error here, though I suggested otherwise in the last commit
message, was legitimate. In particular, the find_match function
should not have searched both windowed and perpetual stores but
one or the other, depending on the AggregationMethod of the
input telemetry. That's because it's possible the model will have
made an aggregation with the same named metric of two distinct
aggregations.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce retainers for windowed telemetry

When we drop telemetry out of a window we have to keep track of the
historical count and sum of those dropped telemetry. This commit
introduces 'retainers' in the prometheus aggregation to do just
that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Perform 'cargo update'

This update brings in a host of smallish updates but the one
we're after is the release of quantiles 0.6.1, fixing the AddAssign
issues with Histogram noticed by @benley in his testing.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct write_text in the prometheus sink

We neglected to do a running sum of the bin values in the text
mode of cernan's prometheus sink. This worked fine in the binary
report.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove fixed-buffer limitation in statsd sink

There's no good reason for us to limit the size that a statsd
payload can be. OS limits will exist and, besides, you probably
should not run statsd on the open internet anyway.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Implement HISTOGRAM support in console sink

This commit does just what it says on the tin, implements support
for HISTOGRAM in the console sink. The implementation is very much
like that in the wavefront sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Create an adjustable, zeroed vector for statsd

This commit walks us back to the vector but makes sure to
have it primed with zeros to avoid upsetting recv_from, which
needs not just the preallocated space but also a primed space.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>