Elasticsearch improvements (#323)
* Tinkering with the ES sink

I'm not seeing the internal metrics I expect out of this sink,
a fact which concerns me. That's the focus of these next few
updates to this sink and related infrastructure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove flush loop, re-create client on each flush

If we fail to make one flush we assume that the client is ruined
and destroy it. In that case, there's no need to loop up to
attempt times in the ES flush. Also, because AWS ES will tend
to hang up if it's not contacted often enough we need to create
a new client each time anyhow.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce self-telemetry around ES errors

The elastic client gives us a pretty good view into why a request
has failed and how. This commit adds logging and telemetry to
expose this information to ops folks.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>