Resolve Prometheus backup issue
In high cardinality situations we observed that the prometheus sink
could back up into its queue. This was a result of the valve_state
being in place for a sink for which it is not appropriate. Sink
telemetry was misleading as the self-telemetry is only updated
during an insertion, which could not happen once valve was locked
shut. The valve state would never clear because prometheus does
not drop timeseries.

This sink also resolves a drop issue with the sink run loop in
some cases and removes the valve state concept from the filter as
that is only complicated and has no real value.

Cernan should now be more robust to high-cardinality timeseries
but the prometheus sink is susceptible to memory attacks if
random names are pumped through.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>