Introducing chan-signal & Signal Handling (#333)
The first step towards supporting at least once delivery sinks, requires
cernan to be able to shutdown gracefully.   This change removes the
inconsistent blocking on child threads in favor of blocking on either
SIGINT or SIGTERM.

Given the state of signal handling of Rust at the time of this writing,
blocking on signals brings with it a dependency on nix and thus limits
Cernan to Unix systems.