Begin the parser for prometheus text protocol
This commit is the first in a series for parsing the prometheus
text protocol. We are now -- mostly -- able to make prometheus
payloads but don't do any parsing of them. Also, we've punted on
building summaries and histograms. The histogram should be easy
to accomidate as all the information we need for producing a
cernan histogram is there. Summary is... harder. I'm not sure
what to do about that as yet.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>