Prometheus freeze up (#340)
* Work-in-progress investigation

The prometheus sink has odd behaviour under high load where
it pulls points off its queue slower over time. This is unique
to the prometheus sink and we now have a minimal testcase for this.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Improve write performance of the prometheus sink

This commit pulls in improvements to both hopper and CKMS. The
changes are:

  * CKMS :: slighly improved write performance, about 20%
  * hopper :: significantly reduced disk write size, about 50%

The reads off the hopper queues into sinks have been made faster
by removing the slow delay. We're still going to have troubles
if the Summarize gets filled up in the prometheus sink: CKMS
write performance degrades sub-linearly to the number of points
stored in the CKMS.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Integrate clippy feedback, test corrections

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Use published hopper and quantiles

The changes that we previously had pinned in branches are now
incorporated into the mainline releases. This commit moves us to
those.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>