Windowing for prometheus sink's Summaries (#344)
* Windowing for prometheus sink's Summaries

This commit re-enables prometheus sink windowing of Summaries but
does so in a way that is more reliable than the old by-total-number-of-points
way we used to do. Operators can now set time limits for windowing.

We no longer allow prometheus sink aggregations to change kind. This
was unlikely do be done in practice but would not work with Prometheus'
expectations and played hell with our quickcheck tests anyhow.

Resolves #343

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct prometheus samples blow-out

Turns out, I goofed the windowing implementation and we'd happily
collect more than the time-limited number of samples. This is now
corrected and there's a quickcheck test to demonstrate it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>