Copy prometheus aggr into HTTP land
The prometheus sink can be made to back up if the prometheus server
holds on to the far-side because we lock the aggregation between
the ingestor and the http server. That means, we have lock contention
that freezes up the ingestion. What we do is copy the ingestion
aggregation every flush interval _when the lock is not held_.
This will mean that some updates will fail to get all the newly
ingested points but we don't hang like dopes waiting for prometheus
server to let go of the request.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>