Remove ES loglines when they've signaled as OK
It's possible for a bulk request to roll through but then for an
individual item in the request to still fail. We now store a UUID
in the ES buffer and use that UUID -- fed through to ES -- to remove
a newly stored bulk item from our buffer.

This resolves #348

Signed-off-by: Brian L. Troutwine <blt@postmates.com>