Remove ES loglines when they've signaled as OK (#349)
* Remove ES loglines when they've signaled as OK

It's possible for a bulk request to roll through but then for an
individual item in the request to still fail. We now store a UUID
in the ES buffer and use that UUID -- fed through to ES -- to remove
a newly stored bulk item from our buffer.

This resolves #348

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Allow elasticsearch sink to shed erroring lines

This commit admits the ability to shed lines that repeatedly error,
letting Ops folks dump these after a configurable limit. This avoids
filling up the memory of Elasticsearch sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clean up match branching

We don't need the Elasticsearch sink to fail if an item cannot
be removed from the internal buffer. This commit clears up that
branch.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>