Bump to 0.8.3
This release bundles some small housekeeping -- updating of deps --
with a fix to the Elasticsearch sink. It turns out that in the case
of a bulk update succeeding ES will still signal that a few lines
have failed. We used to dump those lines. Now, we store them and
re-attempt to submit them up to some configurable limit.

  * 3bc5a255 :: Remove ES loglines when they've signaled as OK (#349)
  * 1d32c5e9 :: Update dependencies via 'cargo update'

Signed-off-by: Brian L. Troutwine <blt@postmates.com>