Correct prometheus window summary truncation (#352)
* Correct prometheus window summary truncation

Turns out, we were truncating the wrong side of the window summary
aggregation. This is now resolved by removing the 0th index from the
summary until we hit our cap.

Resolves #351

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update to use drain

In discussion with @dparton it turns out that the iterator approach
gets a slight advantage over the remove loop in nightly. The code
is also a touch cleaner. Win win.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>