Introduce Telemetry override for 'count' and 'sample_sum' (#354)
* Introduce Telemetry override for 'count' and 'sample_sum'

In some rare cases -- like in the prometheus sink -- it's necessary
to fake the count and sample_sum of a telemetry. Say, if you're
windowing a certain troublesome kind of aggregation. Well, that's
what this commit does: it lets you fake those two things.

I expect the tests are broken now. I'll work those tomorrow.

This work via feedback from @dparton.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Catch the tests up to Prometheus sink changes

This commit catches our tests up to the changes made to Telemetry,
in particular with regard to the overriding of sum and count
for some kinds of Telemetry.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>