Target quantiles 0.7 (#365)
* Target quantiles 0.7

Quantiles 0.7 greatly improved the performance of the CKMS structure
that underpins cernan Summarize. We expect this will resolve issues
with the Prometheus sink when window sizes are non-trivial or the
number of points flowing into cernan is non-trivial.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Re-adjust age_threshold setting in prometheus sink

Turns out the age_threshold hidden feature was not fully wired up.
This is now corrected.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Keep track of exposition delay in Prometheus sink

This commit adds an internal bit of telemetry to report, in micro-
seconds, how long it takes to generate a prometheus exposition. This'll
let Operation keep track of potential issues with the sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>