Introducing Graceful Shutdown (#339)
* Introducing Shutdown Event Type

Shutdown is emitted by the main thread of execution in response to
either SIGTERM or SIGINT.

* Introducing mio Pollers to Shutdown Sources

* Statsd, Internal converted to mio.

Graphite is in progress.  Pushing changes for help/feedback.

* Missed adding constants.rs.

* Resolve capture Fn issue

The primary problem was the Fn may call the external value multiple
times. Happily, we could use an FnOnce here -- since we only call
the function once after making the thread -- and so could move
the outer variables into our new, thready scope.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Graphite Source Ported to Mio

* File Server Source Ported to Mio

* Native Source Ported to Mio

* Introducing cernan::thread.

A new abstraction serving mio enbabled threads.

* Cleanup Prior to Clippy / Rustfmt

* Fixes Derived from Clippy Feedback

* Fixes from Rebasing to Latest Master

* Removed Unused Extern in Util

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Integrate slab into statsd source

The use of HashMap to store token -> socket mapping is pretty
pokey and chews through CPU. This commit integrates slab, a block
allocation library. We use 0.4.0 even though it's a little awkward
because Index is pegged to usize rather than an arbitrary I like
in 0.3.0.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Pin down the shutdown hang issue

I've added a 'shutdown' to the sink to allow for the controlled
shutdown of internal resources. The shutdown will hang for prometheus
sink because of the issue linked in the comments.

We're probably going to have to update prometheus sink HTTP server.
This also implies that, going forward, we'll have to be very careful
about testing shutdown functions for fully running sinks.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Re-introduce shutdown in the statsd source

I misunderstood what John was doing and broke his work. This commit
fixes the thing I broke: statsd source shutdown. What I'd misunderstood
was that the constants::SYSTEM token is _only_ readable when it's time
for a source to go the way of tears in the rain.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Checkpoint on Prometheus Sink Rewrite

* Quiet up the compiler during build

This branch was very chatty during build. It is no longer.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Cargo fmt run

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Re-introduce Sync + Send back into Handler constraints

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Re-introduce 'static for Handler bound in Server::new

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Quieten the build further

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clear out dead comment

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make Sink::shutdown own its structure

Previously shutdown took `&mut self` which meant that the Sink
storage could still be valid after shutdown. This stank in terms
of actually performing shutdown. Now the Sinks can be sure they
own everything on the way out.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Missing Docs Fixes

* Re-introduce write_text back into prometheus sink

This commit re-enables write_text in the prometheus sink. Had
to fiddle with the imports a touch but we're back on the road.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Pass a writable buffer into write_text

write_text no longer creates an http::Response but does require
a buffer to be passed into it, which it then fills up.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Compiling Cleanly.

* Cargo fmt run

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Some clippy suggestions integrated

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Fiddle with references

This commit moves some of the clones we were doing into reference
passes, saving on a handful of copies / increments of atomic counters.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Fix rebase issues

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Convert panic to warn

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove excess externs / marked in review

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Added TokenSlab

Replaces usage of HashMap to store Token -> Evented mappings.

* Feedback from Review #3

* Adding f64 Back to Prometheus Sink

Needed in latest master.

* Remove one extern crate mio

Signed-off-by: Brian L. Troutwine <blt@postmates.com>